backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
supabase>=2.3.0
python-dotenv>=1.0.0
pydantic>=2.3.0,<3.0.0
pydantic-settings>=2.1.0
python-multipart==0.0.6
cryptography==39.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.3.2
aiofiles==23.2.1
